Separate login and signup, add routing
