+++ package.json
@@ -20,6 +20,7 @@
     "vite": "^3.1.0"
   },
   "dependencies": {
-    "@supabase/supabase-js": "^2.0.0-rc.10"
+    "@supabase/supabase-js": "^2.0.0-rc.10",
+    "svelte-routing": "^1.6.0"
   }
 }
