+++ src/App.svelte
@@ -6,6 +6,8 @@
   import Auth from './lib/Auth.svelte'
   import Dashboard from './lib/Dashboard.svelte'
   import { title } from "./stores/title.js";
+  import { Router, Route } from 'svelte-routing';
+  import Signup from './lib/Signup.svelte'
 
   title.clear()
   let session: AuthSession
@@ -19,17 +21,24 @@
       session = _session
     })
   })
+
+  export let url = "";
 </script>
 
 <svelte:head>
 	<title>{$title}</title>
 </svelte:head>
 
-<main class="container" style="padding: 50px 0 100px 0">
-  {#if !session}
-    <Auth />
-  {:else}
-    <Account {session} />
-    <Dashboard />
-  {/if}
-</main>
+<Router url={url}>
+  <main class="container" style="padding: 50px 0 100px 0">
+    {#if !session}
+      <Route path="/"><Auth /></Route>
+      <Route path="/signup"><Signup /></Route>
+    {:else}
+      <Route path="/">
+        <Account {session} />
+        <Dashboard />
+      </Route>
+    {/if}
+  </main>
+</Router>
